# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BusinessNow PSA Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all client accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PROJECTS ─────────────────────────────────────────────────
...
  /allocations:
    get:
      operationId: listAllocations
      tags: [resources]
      summary: List resource allocations
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Allocation"
    post:
      operationId: createAllocation
      tags: [resources]
      summary: Create resource allocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationBody"
      responses:
        "201":
          description: Created allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
  /allocations/{id}:
    patch:
      operationId: updateAllocation
      tags: [resources]
      summary: Update allocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAllocationBody"
      responses:
        "200":
          description: Updated allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
    delete:
      operationId: deleteAllocation
      tags: [resources]
      summary: Delete allocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tax-codes:
    get:
      summary: List tax codes
      operationId: listTaxCodes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaxCode'
    post:
      summary: Create tax code
      operationId: createTaxCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxCodeBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxCode'
  /tax-codes/{id}:
    put:
      summary: Update tax code
      operationId: updateTaxCode
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxCodeBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxCode'
    delete:
      summary: Delete tax code
      operationId: deleteTaxCode
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
  /time-categories:
    get:
      summary: List time categories
      operationId: listTimeCategories
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeCategory'
    post:
      summary: Create time category
      operationId: createTimeCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeCategoryBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeCategory'
  /time-categories/{id}:
    put:
      summary: Update time category
      operationId: updateTimeCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeCategoryBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeCategory'
    delete:
      summary: Delete time category
      operationId: deleteTimeCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: string
        tier:
          type: string
        region:
          type: string
        status:
          type: string
        contractValue:
          type: number
        billingAddress:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        convertedFromProspectId:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - domain
        - tier
        - region
        - status
        - contractValue
        - createdAt
        - updatedAt
    CreateAccountBody:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        tier:
          type: string
        region:
          type: string
        status:
          type: string
        contractValue:
          type: number
        billingAddress:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
      required:
        - name
        - domain
        - tier
        - region
        - contractValue
    UpdateAccountBody:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        tier:
          type: string
        region:
          type: string
        status:
          type: string
        contractValue:
          type: number
        billingAddress:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true

    # ─── PROJECT ────────────────────────────────────────────────
...
    Allocation:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        hoursPerWeek:
          type: number
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - projectId
        - userId
        - startDate
        - endDate
        - hoursPerWeek
        - role
        - createdAt
    CreateAllocationBody:
      type: object
      properties:
        projectId:
          type: integer
        userId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        hoursPerWeek:
          type: number
        role:
          type: string
      required:
        - projectId
        - userId
        - startDate
        - endDate
        - hoursPerWeek
        - role
    UpdateAllocationBody:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        hoursPerWeek:
          type: number
        role:
          type: string
...
    TaxCode:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        rate:
          type: number
        description:
          type: string
          nullable: true
        isDefault:
          type: boolean
        isActive:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - rate
        - isDefault
        - isActive
        - createdAt
    CreateTaxCodeBody:
      type: object
      properties:
        name:
          type: string
        rate:
          type: number
        description:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: integer
      required:
        - name
        - rate
    TimeCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - isActive
        - createdAt
    CreateTimeCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: integer
      required:
        - name
...